Changes: Replaced proxy config with base url variable Added double response request Little refactored code
