+++ angular.json
@@ -63,9 +63,6 @@
         },
         "serve": {
           "builder": "@angular-devkit/build-angular:dev-server",
-          "options": {
-            "proxyConfig": "src/proxy.conf.js"
-          },
           "configurations": {
             "production": {
               "browserTarget": "movie-app:build:production"
