+++ src/app/movie/components/trending-movies/trending-movies.component.ts
@@ -1,22 +1,19 @@
 import {Component, OnInit} from "@angular/core";
 import {Observable} from "rxjs";
 import {MoviesService} from "../../../services/movies.service";
-import {MoviesType} from "../../../types/movies-type";
+import {MoviesType} from "../../../types/movies/movies-type";
 
 @Component({
   selector: "app-trending-movies",
   templateUrl: "./trending-movies.component.html",
   styleUrls: ["./trending-movies.component.scss"]
 })
-export class TrendingMoviesComponent implements OnInit {
+export class TrendingMoviesComponent {
 
   public trendingMovies: Observable<MoviesType[]> = this.moviesService.requestTrendingMovies();
 
+
   constructor(private moviesService: MoviesService) {
-  }
-
-  ngOnInit(): void {
-
   }
 
 }
