+++ src/app/movie/pages/movie-inner-page/components/movie-actors/movie-actors.component.ts
@@ -1,12 +1,12 @@
 import {Component, Input} from '@angular/core';
-import {castCrewType} from "../../../../../types/cast-crew-type";
+import {castType} from "../../../../../types/people/cast-type";
 
 @Component({
-    selector: 'app-movie-actors',
-    templateUrl: './movie-actors.component.html',
-    styleUrls: ['./movie-actors.component.scss']
+  selector: 'app-movie-actors',
+  templateUrl: './movie-actors.component.html',
+  styleUrls: ['./movie-actors.component.scss']
 })
 export class MovieActorsComponent {
 
-    @Input() public castCrew!: castCrewType;
+  @Input() public castActors!: castType;
 }
