+++ src/app/movie/pages/movie-inner-page/components/movie-crew/movie-crew.component.ts
@@ -1,11 +1,11 @@
 import {Component, Input} from '@angular/core';
-import {castCrewType} from "../../../../../types/cast-crew-type";
+import {crewType} from "../../../../../types/people/crew-type";
 
 @Component({
-    selector: 'app-movie-crew',
-    templateUrl: './movie-crew.component.html',
-    styleUrls: ['./movie-crew.component.scss']
+  selector: 'app-movie-crew',
+  templateUrl: './movie-crew.component.html',
+  styleUrls: ['./movie-crew.component.scss']
 })
 export class MovieCrewComponent {
-    @Input() public castCrew!: castCrewType;
+  @Input() public castCrew!: crewType;
 }
