+++ src/app/movie/pages/movie-inner-page/components/movie-dashboard/movie-dashboard.component.ts
@@ -1,12 +1,11 @@
-import {Input} from '@angular/core';
-import {Component} from '@angular/core';
-import {movieOrSeriesType} from "../../../../../types/movies-type";
+import {Component, Input} from '@angular/core';
+import {movieOrSeriesType} from "../../../../../types/movies/movies-type";
 
 @Component({
-    selector: 'app-movie-dashboard',
-    templateUrl: './movie-dashboard.component.html',
-    styleUrls: ['./movie-dashboard.component.scss']
+  selector: 'app-movie-dashboard',
+  templateUrl: './movie-dashboard.component.html',
+  styleUrls: ['./movie-dashboard.component.scss']
 })
 export class MovieDashboardComponent {
-    @Input() public movie!: movieOrSeriesType;
+  @Input() public movie!: movieOrSeriesType;
 }
