+++ src/app/movie/pages/movie-inner-page/movie-inner-page.component.ts
@@ -2,10 +2,8 @@ import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute, Params, Router} from '@angular/router';
 import {combineLatest, Observable} from 'rxjs';
 import {tap} from 'rxjs/operators';
-import {movieOrSeriesType} from '../../../types/movies-type';
-import {castCrewType} from '../../../types/cast-crew-type';
-import {CastCrewService} from '../../../services/cast-crew.service';
 import {MoviesService} from "../../../services/movies.service";
+import {movieOrSeriesType} from "../../../types/movies/movies-type";
 
 @Component({
   selector: 'app-movie-inner-page',
@@ -15,12 +13,10 @@ import {MoviesService} from "../../../services/movies.service";
 export class MovieInnerPageComponent implements OnInit {
 
   public movie$!: Observable<movieOrSeriesType>;
-  public castCrew$!: Observable<castCrewType>;
-  public type: any;
+  public type!: string;
   private paramsId!: string;
 
   constructor(private moviesService: MoviesService,
-              private castCrewService: CastCrewService,
               private readonly router: Router,
               private readonly route: ActivatedRoute
   ) {
@@ -34,9 +30,6 @@ export class MovieInnerPageComponent implements OnInit {
       ).subscribe();
 
     this.movie$ = this.moviesService.requestMovie(this.type, this.paramsId);
-
-    this.castCrew$ = this.castCrewService.requestCast(this.type, this.paramsId);
-
   }
