+++ src/app/services/movies.service.ts
@@ -3,9 +3,9 @@ import {
 } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
+import {map, tap} from 'rxjs/operators';
 import { environment } from '../../environments/environment';
-import {movieOrSeriesType, MoviesType} from '../types/movies-type';
+import {movieOrSeriesType, MoviesType} from "../types/movies/movies-type";
 
 @Injectable({
   providedIn: 'root'
@@ -15,15 +15,15 @@ export class MoviesService {
   constructor(private readonly http: HttpClient) { }
 
   public requestTrendingMovies(time: string = 'week'): Observable<MoviesType[]> {
-    return this.http.get<any>(`/api/trending/all/${time}`).pipe(
+    return this.http.get<any>(`${environment.BASE_URL}/trending/all/${time}`).pipe(
       map(
         response => response.results
-      )
+      ),
     );
   }
 
   public requestMovie(type: string, movieId: string): Observable<movieOrSeriesType> {
-    return this.http.get<movieOrSeriesType>(`/api/${type}/${movieId}`).pipe(
+    return this.http.get<movieOrSeriesType>(`${environment.BASE_URL}/${type}/${movieId}?append_to_response=credits`).pipe(
       map(
         response => response
       )
