+++ src/app/types/enviroment-type.ts
@@ -1,0 +1,4 @@
+export type environmentType = {
+  production: boolean,
+  BASE_URL: string
+}
