--- src/app/types/movies-type.ts
@@ -1,71 +1,0 @@
-export type MoviesType = {
-    adult: boolean,
-    backdrop_path: string,
-    first_air_date?: string,
-    release_date?: string,
-    genre_ids: object,
-    id: number,
-    media_type: string,
-    name?: string,
-    title?: string,
-    origin_country: object,
-    original_language: string,
-    original_name: string,
-    overview: string,
-    popularity: number,
-    poster_path: string,
-    vote_average: number,
-    vote_count: number
-};
-
-export type movieOrSeriesType = {
-    adult: boolean,
-    backdrop_path: string,
-    created_by: [{ id: number, credit_id: string, name: string, gender: number, profile_path: string}],
-    episode_run_time: [number],
-    first_air_date?: string,
-    release_date?: string,
-    genres: [{ id: number, name: string }],
-    homepage: string,
-    id: number,
-    in_production: boolean,
-    languages: [string],
-    last_air_date: string,
-    last_episode_to_air: {
-        air_date: string,
-        episode_number: number,
-        id: number,
-        name: string,
-        overview: string,
-        production_code: string,
-        runtime: number,
-        season_number: number
-        show_id: number,
-        still_path: string,
-        vote_average: number,
-        vote_count: number
-    },
-    name?: string,
-    title?: string,
-    networks: [{ id: number, logo_path: string, name: string, origin_country: string }]
-    next_episode_to_air: string,
-    number_of_episodes: number,
-    number_of_seasons: number,
-    origin_country: [string],
-    original_language: string,
-    original_name: string,
-    overview: string,
-    popularity: number,
-    poster_path: string,
-    production_companies: [{ id: number, logo_path: string, name: string, origin_country: string}],
-    production_countries: [{ iso_3166_1: string, name: string }],
-    seasons: [
-        { air_date: string, episode_count: number, id: number, name: string, overview: string, poster_path: string, season_number: number }
-    ],
-    spoken_languages: [{ english_name: string, iso_639_1: string, name: string }],
-    status: string,
-    tagline: string,
-    type: string,
-    vote_average: number,
-    vote_count: number
-};
