+++ src/app/types/people/cast-crew-type.ts
@@ -1,0 +1,7 @@
+import {castType} from "./cast-type";
+import {crewType} from "./crew-type";
+
+export type castCrewType = {
+  cast: castType,
+  crew: crewType
+};
