+++ src/app/types/people/cast-type.ts
@@ -1,0 +1,8 @@
+export type castType = [{
+  adult: boolean,
+  gender: number,
+  id: number,
+  name: string,
+  character: string,
+  profile_path: string
+}]
