+++ src/app/types/people/crew-type.ts
@@ -1,0 +1,9 @@
+export type crewType = [{
+  adult: boolean,
+  gender: number,
+  id: number,
+  name: string,
+  profile_path: string,
+  department: string,
+  job: string
+}]
