+++ src/environments/environment.prod.ts
@@ -1,3 +1,6 @@
-export const environment = {
-  production: true
+import {environmentType} from "../app/types/enviroment-type";
+
+export const environment: environmentType = {
+  production: true,
+  BASE_URL: 'https://api.themoviedb.org/3'
 };
