+++ src/environments/environment.ts
@@ -2,8 +2,11 @@
 // `ng build --prod` replaces `environment.ts` with `environment.prod.ts`.
 // The list of file replacements can be found in `angular.json`.
 
-export const environment = {
-  production: false
+import {environmentType} from "../app/types/enviroment-type";
+
+export const environment: environmentType = {
+  production: false,
+  BASE_URL: 'https://api.themoviedb.org/3'
 };
 
 /*
